+++ src/common/generic-modal/generic-modal.tsx
@@ -1,5 +1,6 @@
-import type { EventEmitter, VNode} from '@stencil/core';
-import { Component, Event, h,Prop } from '@stencil/core';
+import { Component, Event, h, Prop } from '@stencil/core';
+
+import type { IGenericModalProps } from './generic-modal.types';
 
 @Component({
   tag: 'generic-modal',
@@ -7,17 +8,27 @@ import { Component, Event, h,Prop } from '@stencil/core';
   shadow: false,
 })
 export class GenericModal {
-  @Prop() body: VNode;
-  @Prop() modalTitle: string | VNode;
-  @Prop() modalSubtitle?: string | VNode;
-  @Event() close: EventEmitter<void>;
+  @Prop() body: IGenericModalProps['body'];
+  @Prop() modalTitle: IGenericModalProps['modalTitle'];
+  @Prop() modalSubtitle?: IGenericModalProps['modalSubtitle'];
+  @Event() close: IGenericModalProps['close'];
+
+  private handleClose() {
+    this.close.emit();
+  }
+
+  private handleKeyDown(event: KeyboardEvent) {
+    if (event.key === 'Escape') {
+      this.handleClose();
+    }
+  }
 
   render() {
     return (
-      <div class="modal">
+      <div class="modal" onKeyDown={this.handleKeyDown.bind(this)}>
         <div class="modal-content">
           <div class="modal-header">
-            <span class="close" onClick={() => this.close.emit()}>
+            <span class="close" onClick={this.handleClose.bind(this)}>
               ✕
             </span>
             <h2>{this.modalTitle}</h2>